clean: Change package name
