+++ app/build.gradle.kts
@@ -5,11 +5,11 @@ plugins {
 }
 
 android {
-    namespace = "com.example.expensiarmus"
+    namespace = "com.spellshare.expensiarmus"
     compileSdk = 34
 
     defaultConfig {
-        applicationId = "com.example.expensiarmus"
+        applicationId = "com.spellshare.expensiarmus"
         minSdk = 24
         targetSdk = 34
         versionCode = 1
