+++ app/google-services.json
@@ -9,7 +9,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:793908367016:android:29fde84bc1317eebbb3c9d",
         "android_client_info": {
-          "package_name": "com.example.expensiarmus"
+          "package_name": "com.spellshare.expensiarmus"
         }
       },
       "oauth_client": [],
