--- app/src/main/java/com/example/expensiarmus/data/Balance.kt
@@ -1,6 +1,0 @@
-package com.example.expensiarmus.data
-
-data class Balance(
-    val userUid: String,
-    val amount: Double
-)
