--- app/src/main/java/com/example/expensiarmus/data/Debt.kt
@@ -1,3 +1,0 @@
-package com.example.expensiarmus.data
-
-data class Debt(val debtorUid: String, val creditorUid: String, val amount: Double)
