--- app/src/main/java/com/example/expensiarmus/data/identifiers/ExpenseIdentifier.kt
@@ -1,3 +1,0 @@
-package com.example.expensiarmus.data.identifiers
-
-data class ExpenseIdentifier(override val uid: String) : IIdentifier
