--- app/src/main/java/com/example/expensiarmus/data/identifiers/GroupIdentifier.kt
@@ -1,3 +1,0 @@
-package com.example.expensiarmus.data.identifiers
-
-data class GroupIdentifier(override val uid: String): IIdentifier
