--- app/src/main/java/com/example/expensiarmus/data/identifiers/IIdentifier.kt
@@ -1,5 +1,0 @@
-package com.example.expensiarmus.data.identifiers
-
-interface IIdentifier {
-    val uid: String
-}
