--- app/src/main/java/com/example/expensiarmus/data/identifiers/UserIdentifier.kt
@@ -1,3 +1,0 @@
-package com.example.expensiarmus.data.identifiers
-
-data class UserIdentifier(override val uid: String): IIdentifier
