+++ app/src/main/java/com/spellshare/expensiarmus/data/Balance.kt
@@ -1,0 +1,6 @@
+package com.spellshare.expensiarmus.data
+
+data class Balance(
+    val userUid: String,
+    val amount: Double
+)
