+++ app/src/main/java/com/spellshare/expensiarmus/data/Debt.kt
@@ -1,0 +1,3 @@
+package com.spellshare.expensiarmus.data
+
+data class Debt(val debtorUid: String, val creditorUid: String, val amount: Double)
