+++ app/src/main/java/com/spellshare/expensiarmus/data/identifiers/ExpenseIdentifier.kt
@@ -1,0 +1,3 @@
+package com.spellshare.expensiarmus.data.identifiers
+
+data class ExpenseIdentifier(override val uid: String) : IIdentifier
