+++ app/src/main/java/com/spellshare/expensiarmus/data/identifiers/GroupIdentifier.kt
@@ -1,0 +1,3 @@
+package com.spellshare.expensiarmus.data.identifiers
+
+data class GroupIdentifier(override val uid: String): IIdentifier
