+++ app/src/main/java/com/spellshare/expensiarmus/data/identifiers/IIdentifier.kt
@@ -1,0 +1,5 @@
+package com.spellshare.expensiarmus.data.identifiers
+
+interface IIdentifier {
+    val uid: String
+}
