+++ app/src/main/java/com/spellshare/expensiarmus/data/identifiers/UserIdentifier.kt
@@ -1,0 +1,3 @@
+package com.spellshare.expensiarmus.data.identifiers
+
+data class UserIdentifier(override val uid: String): IIdentifier
